# connector/apiDefinition.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Power BI MCP Bridge",
    "description": "Copilot Studio에서 Power BI Desktop 모델에 접근하는 MCP Bridge. 로컬 Power BI Desktop에 자동 연결을 지원하므로 별도의 연결 설정 없이 바로 도구(table_operations, dax_query_operations 등)를 사용할 수 있습니다.",
    "version": "1.0.0",
    "contact": {
      "name": "Dream I System",
      "url": "https://github.com/your-org/copilot-powerbi-desktop-mcp-bridge"
    }
  },
  "host": "localhost:5050",
  "basePath": "/",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/mcp/tools/list": {
      "post": {
        "operationId": "ListTools",
        "summary": "MCP 도구 목록 조회",
        "description": "사용 가능한 MCP 도구 목록을 반환합니다. Power BI 모델 작업에 사용할 수 있는 모든 도구와 그 파라미터 정보를 확인할 수 있습니다.",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "type": "string",
            "description": "API Key (Bridge 설정에서 활성화한 경우)",
            "x-ms-summary": "API Key"
          }
        ],
        "responses": {
          "200": {
            "description": "도구 목록",
            "schema": {
              "type": "object",
              "properties": {
                "tools": {
                  "type": "array",
                  "description": "사용 가능한 도구 목록",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "description": "도구 이름"
                      },
                      "description": {
                        "type": "string",
                        "description": "도구 설명"
                      }
                    }
                  }
                }
              }
            }
          },
          "502": { "description": "MCP 서버 미연결" },
          "500": { "description": "내부 서버 오류" }
        }
      }
    },
    "/mcp/tools/call": {
      "post": {
        "operationId": "CallTool",
        "summary": "MCP 도구 실행",
        "description": "지정한 MCP 도구를 실행합니다. Power BI Desktop 자동 연결을 지원하여 별도의 connection_operations 호출 없이도 바로 사용할 수 있습니다. toolName으로 도구를 지정하고 toolArguments로 JSON 문자열 또는 객체 형태의 인자를 전달합니다. 대부분의 도구는 중첩된 request 객체 구조를 요구합니다.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["toolName"],
              "properties": {
                "toolName": {
                  "type": "string",
                  "description": "실행할 MCP 도구 이름 (예: table_operations, dax_query_operations, column_operations, measure_operations, relationship_operations, batch_table_operations, connection_operations)",
                  "x-ms-summary": "도구 이름"
                },
                "toolArguments": {
                  "type": "string",
                  "description": "도구 인자 (JSON 문자열 또는 객체). 중첩된 request 구조를 사용합니다. 예시: table_operations → {\"request\":{\"operation\":\"List\"}} | dax_query_operations → {\"request\":{\"operation\":\"Execute\",\"query\":\"EVALUATE ROW(\\\"Result\\\", 1+1)\"}} | column_operations → {\"request\":{\"operation\":\"List\",\"tableName\":\"Sales\"}} | measure_operations → {\"request\":{\"operation\":\"List\",\"tableName\":\"Sales\"}} | connection_operations → {\"request\":{\"operation\":\"Status\"}}",
                  "x-ms-summary": "도구 인자 (JSON)"
                }
              }
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "type": "string",
            "description": "API Key (Bridge 설정에서 활성화한 경우)",
            "x-ms-summary": "API Key"
          }
        ],
        "responses": {
          "200": {
            "description": "도구 실행 결과",
            "schema": {
              "type": "object",
              "properties": {
                "content": {
                  "type": "array",
                  "description": "실행 결과 컨텐츠",
                  "items": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "description": "컨텐츠 타입 (text 등)"
                      },
                      "text": {
                        "type": "string",
                        "description": "결과 텍스트"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "잘못된 요청" },
          "502": { "description": "MCP 서버 미연결" },
          "504": { "description": "요청 시간 초과" },
          "500": { "description": "내부 서버 오류" }
        }
      }
    },
    "/health": {
      "get": {
        "operationId": "HealthCheck",
        "summary": "Bridge 상태 확인",
        "description": "Bridge 서버와 MCP 연결 상태를 확인합니다",
        "parameters": [],
        "responses": {
          "200": {
            "description": "상태 정보",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "description": "Bridge 상태 (ok, error)"
                },
                "bridge": {
                  "type": "object",
                  "description": "Bridge 정보",
                  "properties": {
                    "version": {
                      "type": "string",
                      "description": "Bridge 버전"
                    },
                    "uptime": {
                      "type": "number",
                      "description": "가동 시간 (초)"
                    }
                  }
                },
                "mcp": {
                  "type": "object",
                  "description": "MCP 서버 연결 상태",
                  "properties": {
                    "connected": {
                      "type": "boolean",
                      "description": "연결 여부"
                    },
                    "exe": {
                      "type": "string",
                      "description": "MCP 실행 파일 경로"
                    },
                    "pid": {
                      "type": "integer",
                      "description": "프로세스 ID"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "definitions": {},
  "securityDefinitions": {
    "API Key": {
      "type": "apiKey",
      "in": "header",
      "name": "X-API-Key"
    }
  },
  "security": [],
  "tags": []
}